fix: Separate integration test mocks to prevent unit test failures
Problem: Integration test mocks were breaking unit tests because both test
suites were using the same setup.js file with conflicting mock implementations.

Solution:
- Created separate integration.setup.js for integration-specific mocks
- Moved EventManager, StateManager, and UIManager mocks to integration.setup.js
- Updated jest.config.js to use integration.setup.js for integration tests only
- Kept base mocks (Three.js, CANNON, DOM) in shared setup.js

Results:
- All 978 unit tests now passing
- All 48 integration tests still passing
- No more mock conflicts between test suites
